+++ java/com.sap.sailing.domain.swisstimingadapter/src/com/sap/sailing/domain/swisstimingadapter/SailMasterMessage.java
@@ -20,5 +20,5 @@ public interface SailMasterMessage {
     
     boolean isEvent();
 
-    Integer getSequenceNumber();
+    Long getSequenceNumber();
 }
